+++ tobs_blaine/fxmanifest.lua
@@ -2,8 +2,8 @@ fx_version "cerulean"
 game "gta5"
 
 author "Vega"
-description "Paleto Bay bank heist for ESX"
-version "1.3.1"
+description "Paleto and Fleeca bank heists for ESX"
+version "1.4.0"
 
 shared_scripts {"config/config.lua", "locales/locales.lua"}
 client_scripts {"client/bridge.lua", "client/main.lua"}